set alice as admin
