+++ tee-worker/ts-tests/integration-tests/common/transactions.ts
@@ -1,6 +1,14 @@
+import { hexToU8a } from '@polkadot/util';
 import type { IntegrationTestContext } from './common-types';
 
-import { ApiPromise, FrameSystemEventRecord } from 'parachain-api';
+import {
+    AddressOrPair,
+    ApiPromise,
+    ApiTypes,
+    FrameSystemEventRecord,
+    Keyring,
+    SubmittableExtrinsic,
+} from 'parachain-api';
 
 // for DI-test
 export const subscribeToEventsWithExtHash = async (
@@ -98,4 +106,60 @@ export async function waitForBlock(api: ApiPromise, blockNumber: number, blocksT
             }
         });
     });
+}
+
+async function setAliceAsAdmin(api: ApiPromise) {
+    // Get keyring of Alice, who is also the sudo in dev chain spec
+    const keyring = new Keyring({ type: 'sr25519' });
+    const alice = keyring.addFromUri('//Alice');
+
+    const tx = await sudoWrapperGC(api, api.tx.teebag.setAdmin('esqZdrqhgH8zy1wqYh1aLKoRyoRWLFbX9M62eKfaTAoK67pJ5'));
+
+    console.log(`Setting Alice as Admin for Teebag`);
+    return signAndSend(tx, alice);
+}
+
+export function signAndSend(tx: SubmittableExtrinsic<ApiTypes>, account: AddressOrPair) {
+    return new Promise<{ block: string }>(async (resolve, reject) => {
+        await tx.signAndSend(account, (result) => {
+            console.log(`Current status is ${result.status}`);
+            if (result.status.isInBlock) {
+                console.log(`Transaction included at blockHash ${result.status.asInBlock}`);
+            } else if (result.status.isFinalized) {
+                console.log(`Transaction finalized at blockHash ${result.status.asFinalized}`);
+                resolve({
+                    block: result.status.asFinalized.toString(),
+                });
+            } else if (result.status.isInvalid) {
+                reject(`Transaction is ${result.status}`);
+            }
+        });
+    });
+}
+
+// After removing the sudo module, we use `EnsureRootOrHalfCouncil` instead of `Sudo`,
+// and there are only two council members in litmus-dev/rococo-dev/litentry-dev.
+// So only `propose` is required, no vote.
+//
+// TODO: support to send the `vote extrinsic`, if the number of council members is greater than 2.
+export async function sudoWrapperGC(api: ApiPromise, tx: SubmittableExtrinsic<ApiTypes>) {
+    const chain = (await api.rpc.system.chain()).toString().toLowerCase();
+    if (chain != 'rococo-dev') {
+        const threshold = api.createType('Compact<u32>', 1);
+        const call = api.createType('Call', tx);
+        return api.tx.council.propose(threshold, call, api.createType('Compact<u32>', tx.length));
+    } else {
+        return api.tx.sudo.sudo(tx);
+    }
+}
+
+export async function setScheduledEnclave(api: ApiPromise, block: number, mrenclave: string) {
+    const keyring = new Keyring({ type: 'sr25519' });
+    const alice = keyring.addFromUri('//Alice');
+
+    await setAliceAsAdmin(api);
+    const tx = api.tx.teebag.setScheduledEnclave('Identity', block, hexToU8a(`0x${mrenclave}`));
+
+    console.log('Schedule Enclave Extrinsic sent');
+    return signAndSend(tx, alice);
 }
